+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - `timeline-suspicious-processes` would crash when Hayabusa results from version 2.8.0+ was used. (#35) (@fukusuket)
 - Fixed a JSON parsing error in VirusTotal lookups when an invalid API key was specified. (@fukusuket)
+- `timeline-suspicious-processes` was not correctly outputting `ParentPGUID` field. Improved PID decimal conversion. (#50) (@fukusuket)
 
 ## 2.0.0 [2022/08/03] - [SANS DFIR Summit 2023 Release](https://www.sans.org/cyber-security-training-events/digital-forensics-summit-2023/)